Move `cuda::barrier` to its own file
